+++ src/imio/helpers/tests/test_converters.py
@@ -2,7 +2,7 @@
 
 from imio.helpers import HAS_PLONE_5_AND_MORE
 from imio.helpers.testing import IntegrationTestCase
-from unittest import makeSuite
+from unittest import TestLoader
 from unittest import TestSuite
 from ZPublisher.tests.testHTTPRequest import HTTPRequestTests
 
@@ -36,6 +36,8 @@ class TestConverters(IntegrationTestCase, HTTPRequestTests):
 
 def test_suite():
     suite = TestSuite()
-    # change prefix to avoid executing every tests of HTTPRequestTests
-    suite.addTest(makeSuite(TestConverters, prefix='test_imiohelpers_'))
+    for name in TestLoader().getTestCaseNames(TestConverters):
+        # change prefix to avoid executing every tests of HTTPRequestTests
+        if name.startswith('test_imiohelpers_'):
+            suite.addTest(TestConverters(name))
     return suite